scripts: resync all build scripts

+++ build_kernel.sh
@@ -45,19 +45,22 @@ copy_defconfig () {
 	cd "${DIR}/KERNEL" || exit
 	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
 
-	#./ti_config_fragments/defconfig_builder.sh -l
-	./ti_config_fragments/defconfig_builder.sh -t ti_sdk_am3x_release
-	./ti_config_fragments/defconfig_builder.sh -t ti_sdk_dra7x_release
+	if [ -f ./ti_config_fragments/defconfig_builder.sh ] ; then
+		#./ti_config_fragments/defconfig_builder.sh -l
+		./ti_config_fragments/defconfig_builder.sh -t ti_sdk_am3x_release
+		./ti_config_fragments/defconfig_builder.sh -t ti_sdk_dra7x_release
 
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_am3x_release_defconfig
-	cp -v .config "${DIR}/patches/ti_sdk_am3x_release_defconfig"
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_am3x_release_defconfig
+		cp -v .config "${DIR}/patches/ti_sdk_am3x_release_defconfig"
 
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_dra7x_release_defconfig
-	cp -v .config "${DIR}/patches/ti_sdk_dra7x_release_defconfig"
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" ti_sdk_dra7x_release_defconfig
+		cp -v .config "${DIR}/patches/ti_sdk_dra7x_release_defconfig"
 
-	make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" distclean
+	fi
+
 	if [ ! -f "${DIR}/.yakbuild" ] ; then
 		make ARCH=${KERNEL_ARCH} CROSS_COMPILE="${CC}" "${config}"
 		cp -v .config "${DIR}/patches/ref_${config}"
